docker/training-worker/requirements.txt
kafka-python-ng==2.2.3
huggingface_hub>=1.0.0
httpx>=0.27.0
